Update: v1 toml header (#195)
* Playing around

* Screwing around

* No csv progress

* Update libraries

* Progress

* Progress

* Fix bunch of testing

* Separated file handler

* Progress

* Thread weight progress

* Yes

* Finished the threads distribution function

* Working GeomModel with new header structure

* Update config checking

* Fix output type

* Writing progress

* Almost working

* Progess

* More progess

* Done

* Update sonarqube call

* Scan issues

* Scan issues #2

+++ src/fiat/fio/buffer.py
@@ -4,13 +4,11 @@ import os
 from io import BytesIO, FileIO
 from multiprocessing.synchronize import Lock
 from pathlib import Path
-from typing import Any
 
-from osgeo import ogr, osr
+from osgeo import gdal, ogr, osr
 
 from fiat.fio.fopen import open_geom
 from fiat.fio.geom import GeomIO
-from fiat.fio.misc import merge_geom_layers
 from fiat.util import (
     NEWLINE_CHAR,
     DummyLock,
@@ -26,10 +24,6 @@ class BufferedGeomWriter:
     ----------
     file : Path | str
         Path to the file.
-    srs : osr.SpatialReference
-        The spatial reference system of the file (and the buffer).
-    layer_defn : ogr.FeatureDefn, optional
-        The definition of the layer, by default None
     buffer_size : int, optional
         The size of the buffer, by default 100000
     lock : Lock, optional
@@ -39,14 +33,12 @@ class BufferedGeomWriter:
     def __init__(
         self,
         file: Path | str,
-        srs: osr.SpatialReference,
-        layer_defn: ogr.FeatureDefn = None,
         buffer_size: int = 100000,  # geometries
         lock: Lock = None,
     ):
         # Ensure pathlib.Path
-        file = Path(file)
-        self.file: Path = file
+        path = Path(file)
+        self.path: Path = path
 
         # Set the lock
         self.lock: Lock | DummyLock = lock
@@ -57,66 +49,74 @@ class BufferedGeomWriter:
         self.pid: int = os.getpid()
 
         # Set for later use
-        self.srs: osr.SpatialReference = srs
-        self.flds: dict[str:int] = {}
         self.n: int = 1
-
-        if layer_defn is None:
-            with open_geom(self.file, mode="r") as _r:
-                layer_defn = _r.layer.defn
-            _r = None
-        self.layer_defn: ogr.FeatureDefn = layer_defn
 
         # Create the buffer
         self.buffer: GeomIO = open_geom(f"/vsimem/{file.stem}.gpkg", mode="w")
-        self.buffer.create_layer(
-            srs,
-            layer_defn.GetGeomType(),
-        )
-        self.buffer.layer.set_from_defn(
-            layer_defn,
-        )
+
         # Set some check vars
         # TODO: do this based om memory foodprint
-        # Needs some reseach into ogr's memory tracking
+        # Needs some reseach into ogr's exposed memory tracking
         self.max_size: int = buffer_size
         self.size: int = 0
 
     def __del__(self):
         self.buffer = None
-        self.layer_defn = None
 
-    def __reduce__(self) -> str | tuple[Any, ...]:
-        pass
+    ## State altering
+    def close(self) -> None:
+        """Close the buffer."""
+        # Flush on last time
+        if self.buffer.layer is not None:
+            self.write(reset=False)
+        self.buffer.delete(all=True)
+        self.buffer.close()
 
-    def _reset_buffer(self):
+    def reset_buffer(self):
+        """Reset the buffer to an empty dataset/ layer."""
+        if self.buffer.layer is None:
+            return
+        # Get the define
+        defn = self.buffer.layer.defn
+        srs = self.buffer.srs
         # Delete
         self.buffer.delete()
 
         # Re-create
-        self.buffer.create_layer(
-            self.srs,
-            self.layer_defn.GetGeomType(),
-        )
-        self.buffer.layer.set_from_defn(
-            self.layer_defn,
-        )
-        self.create_fields(self.flds)
+        self.setup(defn, srs)
 
         # Reset current size
         self.size = 0
+        defn = None
+        srs = None
 
-    def close(self):
-        """Close the buffer."""
-        # Flush on last time
-        self.to_drive()
-        self.buffer.delete(all=True)
-        self.buffer.close()
+    ## I/O
+    def write(self, reset: bool = True) -> None:
+        """Dump the buffer to the drive."""
+        # Block while writing to the drive
+        self.buffer.flush()
+        # self.buffer.src.Close()
 
+        # Lock and merge/ write
+        self.lock.acquire()
+        ds = gdal.VectorTranslate(
+            self.path.as_posix(),
+            self.buffer.path.as_posix(),
+            format="GPKG",
+            accessMode="append",
+        )
+        ds.Close()
+        ds = None
+        self.lock.release()
+
+        if reset:
+            self.reset_buffer()
+
+    ## Mutating methods
     def add_feature(
         self,
         ft: ogr.Feature,
-    ):
+    ) -> None:
         """Add a feature to the buffer.
 
         Parameters
@@ -125,16 +125,16 @@ class BufferedGeomWriter:
             The feature.
         """
         if self.size + 1 > self.max_size:
-            self.to_drive()
+            self.write()
         self.buffer.layer.add_feature(ft)
-
+        # Added a new feature, so plus 1
         self.size += 1
 
     def add_feature_with_map(
         self,
         ft: ogr.Feature,
         fmap: dict,
-    ):
+    ) -> None:
         """Add a feature to the buffer with additional field info.
 
         Parameters
@@ -146,40 +146,29 @@ class BufferedGeomWriter:
 the fields in the buffer.
         """
         if self.size + 1 > self.max_size:
-            self.to_drive()
+            self.write()
         self.buffer.layer.add_feature_with_map(
             ft,
             fmap=fmap,
         )
-
+        # Added a new feature, so plus 1
         self.size += 1
 
-    def create_fields(
+    def setup(
         self,
-        flds: zip,
-    ):
-        """Create new fields in the buffer dataset."""
-        _new = dict(flds)
-        self.flds.update(_new)
+        defn: ogr.FeatureDefn,
+        srs: osr.SpatialReference,
+        extra_fields: dict | zip | None = None,
+    ) -> None:
+        """Set up a layer for the buffer."""
+        # Create the layer
+        self.buffer.create_layer(srs, geom_type=defn.GetGeomType())
+        self.buffer.layer.set_from_defn(defn)
 
-        self.buffer.layer.create_fields(
-            _new,
-        )
-
-    def to_drive(self):
-        """Dump the buffer to the drive."""
-        # Block while writing to the drive
-        # self.buffer.close()
-        self.lock.acquire()
-        merge_geom_layers(
-            self.file.as_posix(),
-            f"/vsimem/{self.file.stem}.gpkg",
-            out_layer_name=self.file.stem,
-        )
-        self.lock.release()
-
-        # self.buffer = self.buffer.reopen(mode="w")
-        self._reset_buffer()
+        # Update the layer with new fields
+        if extra_fields is not None:
+            extra_fields = dict(extra_fields)  # Ensure typing
+            self.buffer.layer.create_fields(extra_fields)
 
 
 class BufferedTextWriter(BytesIO):
@@ -216,16 +205,17 @@ class BufferedTextWriter(BytesIO):
         )
         self.max_size = buffer_size
 
-    def close(self):
+    def close(self) -> None:
         """Close the writer and the buffer."""
         # Flush on last time
-        self.to_drive()
+        self.dump()
         self.stream.close()
 
         # Close the buffer
         BytesIO.close(self)
 
-    def to_drive(self):
+    ## I/O
+    def dump(self) -> None:
         """Dump to buffer to the drive."""
         self.seek(0)
 
@@ -240,10 +230,11 @@ class BufferedTextWriter(BytesIO):
         self.truncate(0)
         self.seek(0)
 
-    def write(
+    ## Mutating methods
+    def add(
         self,
         b: bytes,
-    ):
+    ) -> None:
         """Write bytes to the buffer.
 
         Parameters
@@ -252,14 +243,14 @@ class BufferedTextWriter(BytesIO):
             Bytes to write.
         """
         if self.tell() + len(b) > self.max_size:
-            self.to_drive()
-        BytesIO.write(self, b)
+            self.dump()
+        self.write(b)
 
-    def write_iterable(self, *args):
+    def add_iterable(self, *args) -> None:
         """Write a multiple entries to the buffer."""
         by = b""
         for arg in args:
             by += ("," + "{}," * len(arg)).format(*arg).rstrip(",").encode()
         by = by.lstrip(b",")
         by += NEWLINE_CHAR.encode()
-        self.write(by)
+        self.add(by)
